Update SensingDevice.cpp
No code was present for intialzing input and output pins, so I added it to the member Init of the SensingDevice class. Additionally, in order for the successful retrival of the button states from the controllers during the very first execution of the relevent for loop, the clock pin has to be set LOW in advance. The reason for this is that data is put out by the controllers on the rising edge of the clock signal (transition from low to high).